+++ fn-extra/CHANGELOG.md
@@ -1,7 +1,13 @@
+* 0.?.?.? Daniel Patterson <dbp@dbpmail.net> 2015-11-??
+
+  - Add `heistServe`, which serves templates according to path.
+  - Remove `heistLens` and `setHeist` from `HeistContext` type class,
+    as they aren't used.
+
 * 0.1.1.0 Daniel Patterson <dbp@dbpmail.net> 2015-10-30
 
-  Add `tag'`, which builds splices without atttributes.
+  - Add `tag'`, which builds splices without atttributes.
 
 * 0.1.0.0 Daniel Patterson <dbp@dbpmail.net> 2015-10-26
 
-  Initial release.
+  - Initial release.